Prediction_Diabetes.ipynb
Can you develop a machine learning model that can predict whether people have diabetes or not when features are identified?

The dataset is a large dataset held at the National Institutes of Diabetes-Digestive-Kidney Diseases in the USA. Data used in diabetes studies on Pima Indian women aged 21 and over living in Phoenix, the 5th largest city in the US state of Arizona. It consists of 768 observations and 8 numerical independent variables. The target variable is specified as “result”; 1 indicates a positive diabetes test result and 0 indicates a negative result.